reset animal duel distances to original gladiator duel distances
Using 6~8 causes escort pawns to stand in a circle centered
on the speaker, not on the ritual spot. Using 4~4 for the speaker
makes him covered by the name label of the pawn standing one
cell up. Using 5~5 for the speaker causes again cell reservation
problems like with funeral pyre. Using 3~3 for the speaker,
the normal setting, is visually too close to the fight (it IMO looks
worse than with two pawns), but let's keep it that way for now.