change obsolete rituals back to visible
It seems to cause them to be silently added when adding the mod
(specifically: start without mod, choose fluid ideo, animist+nature,
save the game, add the mod, load the game, see the tooltip for
ideo points, the rituals get listed and pawns will want them).

The C# code disabling adding of those rituals seems to be enough
to also hide them from the UI, so this XML is not needed.