try to use a normal left-to-die calculation for Compassion if possible
The anti-cheese calculation based on recording expected time to die
has the problem that it's an estimate and apparently in some cases
it may be way off. The shuttle incident from Vanilla Events Expanded
seems to cause a long expected time for pawns that die almost
immediately.

Keep the anti-cheese calculation only when the downed pawn is actually
harmed by a colonist outside of a fight.